fix: changed steeringagent, NPC attract, BarkFlee, removed NPCWander & Patrol
So far, changed NPCAttract, BarkFlee, and a little bit of SteeringAgent. NPC scripts override the target position for the steering agent but do not pass any force, hence the NPC does not move towards the bark position and continues any behaviour they were previously doing. Maybe pass in a force as well ? Not sure if this is how we want to do it