feat: better NPCs (ft. Nono)
* feat: add initial steering behaviours

* fix: fixed errors due to name changes

fixed few errors in the project due to name changes in scripts

* feat: new npc prefab for testing

added new npc prefab for testing. Things work, but errors from draggable objects (non-convex meshcollider with non-kinematic rigidbody error)

* fix: changed steeringagent, NPC attract, BarkFlee, removed NPCWander & Patrol

So far, changed NPCAttract, BarkFlee, and a little bit of SteeringAgent. NPC scripts override the target position for the steering agent but do not pass any force, hence the NPC does not move towards the bark position and continues any behaviour they were previously doing. Maybe pass in a force as well ? Not sure if this is how we want to do it

* Revert "fix: changed steeringagent, NPC attract, BarkFlee, removed NPCWander & Patrol"

This reverts commit 6fb779204b378c266c238a4bf4e39931dd82d069.

* fix: Changes to steeringagent to have Rigidbody movement, and BarkFlee for NPCs

added few changes to SteeringAgent to handle movement based on rigidbodies (apply force to rigidbodies to move), BarkFlee for NPC to integrate SteeringBehaviours

* clean: refactor nono's code :P

* fix: reposition players and kill floor

---------

Co-authored-by: dhruvsaikia <85579996+dhruvsaikia@users.noreply.github.com>
Co-authored-by: dhruvsaikia <dhruvsaikia@gmail.com>